requirements.txt
# IGEV ---------------------------------------
scikit-image
tensorboard
matplotlib
tqdm
timm==0.5.4

pytorch==2.1.0
pytorch-cuda==11.7
pytorch-mutex==1.0

torchaudio==2.1.0
cudatoolkit==11.3

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

